data export/import fixes
-refresh dir now has an option to not delete the directory
-shader export now supports selection export/import
-shader export now has a clear shaders button
-animation version import now works
-attribute version import now works
-weight deformer versions now works

+++ python/vtool/util_file.py
@@ -2481,7 +2481,7 @@ def delete_read_only_error(action, name, exc):
     action(name)
     
 
-def refresh_dir(directory):
+def refresh_dir(directory, delete_directory = True):
     """
     Delete everything in the directory.
     """
@@ -2490,16 +2490,18 @@ def refresh_dir(directory):
     dir_name = get_dirname(directory)
     
     if exists(directory):
+        
         try:
-            files = get_files(directory)
+            files = get_files_and_folders(directory)
         except:
             files = []
         
         if files:
             for filename in files:
                 delete_file(filename, directory)
-            
-        delete_dir(base_name, dir_name)
+        
+        if delete_directory:
+            delete_dir(base_name, dir_name)
         
     if not exists(directory):
         create_dir(base_name, dir_name)
